Define keywords for flex components (#46)
This pull request defines the keywords that can be specified for
template messages and flex messages as enums. In addition to keywords,
template messages and flex messages can specify values in percentages or
pixels.
(e.g. image size in flex message:
https://developers.line.biz/en/docs/messaging-api/flex-message-layout/#image-size)

Therefore, we will not define the new enum as a value that can be
specified for template messages and flex messages.

The main purpose of this change is to provide support for users who use
the bot SDK to have the option to use enums.

Resolve https://github.com/line/line-openapi/issues/42