Show links to access generated code easily in PR (#76)
In the `Show diff` step, all jobs display the differences of the
generated code. While some members have navigated by opening the GitHub
Actions job URL, it should be more convenient to comment directly on the
PR.

This change supports that.

(Additional Notes on the Comments)
1. Since the generated code may be updated with each additional commit,
the job deletes old comments containing old job URLs if present.
2. Some generators are incomplete, so the comment shows tasks directed
to the person who updated the OpenAPI YAML, especially when
`webhook.yml` is updated.