Add new errorCode 4 of NarrowcastProgressResponse (#79)
A new errorCode has been added to NarrowcastProgressResponse.
New number is 4. 4 means an audience of less than 50 recipients is
included as a condition of sending.

Previously, it was not possible to determine whether an error occurred
due to a lack of enough recipients or because an audience of less than
50 recipients is included as a condition of sending. This will be fixed
with the update to the Messaging API, so the description of errorCode on
OpenAPI will also be revised.