Warn PR from forked repository (#97)
This repository is public, but contributions are primarily expected from
our employees. Additionally, some automations are disabled for forked
repositories (technically possible, but not implemented).

Even the main contributors sometimes forget not to use forked repos.
Using a forked repo causes CI to fail, but it takes time to remember
this. Therefore, this change starts to remind contributors to push
branches to this repository instead of using forked repos.

(example)

![image](https://github.com/user-attachments/assets/25adfb1d-9787-42df-bd7f-c31d191e75cf)


This change also adds this expectation to the README. Generally,
discussions should be held in issues at first unless it's a really
trivial typo. Even if we have valid reasons, it's disheartening to
reject contributions after someone has put in the effort.