Add type as required property in source object of webhook (#103)
In the Messaging API webhook, we use `discriminator` to differentiate
classes based on the value of `type`. During the development of bot SDK
for Ruby v2, it seems that only `Source` does not have `type` marked as
required. This is simply a mistake. The type should always be included
to distinguish the kind of JSON, and there should not be any instances
where it is absent.