Warn if underscore is contained in JSON for request body (#107)
line-bot-sdk-ruby v2.0.0 automatically converts underscores of JSON's
key in the request body to camel case. Thus, if there are properties
with underscores, the request will fail due to this implementation.
In the first place, many LINE developers' products uses already camel
case in so many features, so there's no reason to use underscores as of
now. (though we know some properties use underscores... we can't revert
it after publishing it)

This change adds a workflow and script that will cause CI to fail if
underscores might be included, reminding us of this fact.