Add forbidPartialDelivery option to the Narrowcast Limit Object (#114)
## Add forbidPartialDelivery option to the Narrowcast Limit Object

We add a new `forbidPartialDelivery` option to the Narrowcast Limit
Object.

When set to true, this option prevents messages from being delivered to
only a subset of the target audience.
If partial delivery occurs, the narrowcast request will succeed but fail
asynchronously.
You can verify whether the message delivery was canceled by checking the
narrowcast message progress.

This property can only be set to true when upToRemainingQuota is also
true.

For more details, see the
https://developers.line.biz/en/news/2025/10/21/narrowcast-message-update/.

### Example:
```json
{
  "max": 100,
  "upToRemainingQuota": true,
  "forbidPartialDelivery": true
}
```