Support mark as read by token API (#115)
## Support for "Mark as Read" by Token API

We have released a new **Mark as Read API** that allows developers to
mark a user’s messages as read.
Previously, this functionality was available only to partners, but it is
now publicly available.

When your server receives a user message via Webhook, the `MessageEvent`
will include a new field: `markAsReadToken`.
By calling the Mark as Read API with this token, all messages in the
chat room **up to and including** that message will be marked as read.

> **Note:** This feature assumes that your service uses the chat feature
through Official Account Manager.
> If chat is not enabled, messages from users are automatically marked
as read, making this API unnecessary.

For more details, please refer to the release note:
https://developers.line.biz/en/news/2025/11/05/mark-as-read/